# pva.SuperV.Api/openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SuperV API",
    "description": "API for accessing SuperV projects.",
    "version": "v1"
  },
  "paths": {
...
    "/classes/{projectId}": {
      "get": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Gets the list of available classes in a project",
        "description": "Gets the list of available classes in a project",
        "operationId": "GetClasses",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Creates a class in a project",
        "description": "Creates a class in a project",
        "operationId": "CreateClass",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Class creation request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/classes/{projectId}/{className}": {
      "get": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Gets a class of a project by its name",
        "description": "Gets a class of a project by its name",
        "operationId": "GetClass",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "className",
            "in": "path",
            "description": "Name of class",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Deletes a class of a project by its name",
        "description": "Deletes a class of a project by its name",
        "operationId": "DeleteClass",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "className",
            "in": "path",
            "description": "Name of class",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/fields/{projectId}/{className}": {
      "get": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Gets the list of available fields in a class from a project",
        "description": "Gets the list of available fields in a class from a project",
        "operationId": "GetFieldDefinitions",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "className",
            "in": "path",
            "description": "Name of class",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldDefinitionModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Creates a field definition in a class of a project",
        "description": "Creates a field definition in a class of a project",
        "operationId": "CreateFieldDefinition",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "className",
            "in": "path",
            "description": "Name of class",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Field creation request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldDefinitionModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDefinitionModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/fields/{projectId}/{className}/{fieldName}": {
      "get": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Gets a field definition from a class of a project by its name",
        "description": "Gets a field definitionfrom a class of a project by its name",
        "operationId": "GetFieldDefinition",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "className",
            "in": "path",
            "description": "Name of class",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fieldName",
            "in": "path",
            "description": "Name of field",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldDefinitionModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pva.SuperV.Api"
        ],
        "summary": "Deletes a field definition from a class of a project by its name",
        "description": "Deletes a field definition from a class of a project by its name",
        "operationId": "DeleteFieldDefinition",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "className",
            "in": "path",
            "description": "Name of class",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fieldName",
            "in": "path",
            "description": "Name of field",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClassModel": {
        "required": [
          "name",
          "baseclassName"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the class."
          },
          "baseclassName": {
            "type": "string",
            "description": "Base class (if any).",
            "nullable": true
          }
        }
      },
...
      "FieldDefinitionModel": {
        "required": [
          "$type"
        ],
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/FieldDefinitionModelBoolFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelDateTimeFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelDoubleFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelFloatFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelIntFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelLongFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelShortFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelStringFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelTimeSpanFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelUintFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelUlongFieldDefinitionModel"
          },
          {
            "$ref": "#/components/schemas/FieldDefinitionModelUshortFieldDefinitionModel"
          }
        ],
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "BoolFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelBoolFieldDefinitionModel",
            "DateTimeFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelDateTimeFieldDefinitionModel",
            "DoubleFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelDoubleFieldDefinitionModel",
            "FloatFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelFloatFieldDefinitionModel",
            "IntFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelIntFieldDefinitionModel",
            "LongFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelLongFieldDefinitionModel",
            "ShortFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelShortFieldDefinitionModel",
            "StringFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelStringFieldDefinitionModel",
            "TimeSpanFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelTimeSpanFieldDefinitionModel",
            "UintFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelUintFieldDefinitionModel",
            "UlongFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelUlongFieldDefinitionModel",
            "UshortFieldDefinitionModel": "#/components/schemas/FieldDefinitionModelUshortFieldDefinitionModel"
          }
        }
      },
      "FieldDefinitionModelBoolFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "BoolFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "boolean",
            "description": "Default value"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelDateTimeFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "DateTimeFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "string",
            "description": "Default value",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelDoubleFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "DoubleFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "number",
            "description": "Default value",
            "format": "double"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelFloatFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "FloatFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "number",
            "description": "Default value",
            "format": "float"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelIntFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "IntFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelLongFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "LongFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelShortFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "ShortFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "int16"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelStringFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "StringFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "string",
            "description": "Default value",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelTimeSpanFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "TimeSpanFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string",
            "description": "Default value"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelUintFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "UintFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "uint32"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelUlongFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "UlongFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "uint64"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
      "FieldDefinitionModelUshortFieldDefinitionModel": {
        "required": [
          "defaultValue",
          "name"
        ],
        "properties": {
          "$type": {
            "enum": [
              "UshortFieldDefinitionModel"
            ],
            "type": "string"
          },
          "defaultValue": {
            "type": "integer",
            "description": "Default value",
            "format": "uint16"
          },
          "name": {
            "type": "string",
            "description": "Name of field"
          },
          "type": {
            "type": "string",
            "description": "Type of field"
          }
        }
      },
...
    }
  },
...
}